+++ power/power-8226.c
@@ -95,7 +95,7 @@ int power_hint_override(__attribute__((unused)) struct power_module *module,
         power_hint_t hint, void *data)
 {
     if (hint == POWER_HINT_SET_PROFILE) {
-        set_power_profile((int)data);
+        set_power_profile((intptr_t)data);
         return HINT_HANDLED;
     }
 
@@ -113,7 +113,7 @@ int power_hint_override(__attribute__((unused)) struct power_module *module,
     }
 
     if (hint == POWER_HINT_CPU_BOOST) {
-        int duration = (int)data / 1000;
+        int duration = (intptr_t)data / 1000;
         int resources[] = { CPUS_ONLINE_MIN_2, 0x20F, 0x30F};
 
         if (duration > 0)
